Fixed the Wakatime stats.
It was stuck showing the wrong stats and it turned out to be to do with me renaming the title of the card. If I decreased the character count by one, it would work, and if I increased it, it would work. So I've expanded 'Stats' into 'Statistics' and now the correct info is being shown.